feat: add loaded values (for edit)
